User can see the movies accordingly as Popular and Top Rated.
When the user clicks on a movie it goes to a details screen where the user can see the details of the movie and also the movie trailers and the movie reviews.The app is in the intermediate stage right now.I'm working on how to favorite a movie so that when the user is offline the user can see the favorite movies. P.S- This project is a part of the Android Developer Nanodegree Program from Udacity in which I am currently enrolled.